Rename execute_path -> execute_trajectory
