+++ moveit_ros/move_group/src/default_capabilities/execute_trajectory_action_capability.cpp
@@ -34,7 +34,7 @@
 
 /* Author: Kentaro Wada */
 
-#include "execute_path_action_capability.h"
+#include "execute_trajectory_action_capability.h"
 
 #include <moveit/plan_execution/plan_execution.h>
 #include <moveit/trajectory_processing/trajectory_tools.h>
